refactor: creates pure extractFirstWord class
